move connector common to a dir
