+++ src/connector/src/lib.rs
@@ -52,8 +52,7 @@ pub mod schema;
 pub mod sink;
 pub mod source;
 
-pub mod common;
-pub mod mqtt_common;
+pub mod connector_common;
 
 pub use paste::paste;
