+++ src/connector/src/parser/util.rs
@@ -21,7 +21,7 @@ use risingwave_common::types::Datum;
 use risingwave_pb::data::DataType as PbDataType;
 
 use crate::aws_utils::load_file_descriptor_from_s3;
-use crate::common::AwsAuthProps;
+use crate::connector_common::AwsAuthProps;
 use crate::error::ConnectorResult;
 use crate::source::SourceMeta;
