+++ src/connector/src/source/filesystem/s3/enumerator.rs
@@ -17,7 +17,7 @@ use async_trait::async_trait;
 use aws_sdk_s3::client::Client;
 
 use crate::aws_utils::{default_conn_config, s3_client};
-use crate::common::AwsAuthProps;
+use crate::connector_common::AwsAuthProps;
 use crate::source::filesystem::file_common::FsSplit;
 use crate::source::filesystem::s3::S3Properties;
 use crate::source::{FsListInner, SourceEnumeratorContextRef, SplitEnumerator};
