+++ src/connector/src/source/filesystem/s3/mod.rs
@@ -20,7 +20,7 @@ mod source;
 use serde::Deserialize;
 pub use source::S3FileReader;
 
-use crate::common::AwsAuthProps;
+use crate::connector_common::AwsAuthProps;
 use crate::source::filesystem::FsSplit;
 use crate::source::{SourceProperties, UnknownFields};
