+++ src/connector/src/source/kafka/mod.rs
@@ -17,7 +17,7 @@ use std::collections::HashMap;
 use serde::Deserialize;
 use serde_with::{serde_as, DisplayFromStr};
 
-use crate::common::KafkaPrivateLinkCommon;
+use crate::connector_common::KafkaPrivateLinkCommon;
 
 pub mod enumerator;
 pub mod private_link;
@@ -31,7 +31,7 @@ pub use source::*;
 pub use split::*;
 use with_options::WithOptions;
 
-use crate::common::{KafkaCommon, RdKafkaPropertiesCommon};
+use crate::connector_common::{KafkaCommon, RdKafkaPropertiesCommon};
 use crate::source::SourceProperties;
 
 pub const KAFKA_CONNECTOR: &str = "kafka";
